requirements.txt
streamlit>=1.28
plotly>=5.18
pandas>=2.0
pyyaml>=6.0
